Add -m/--metadata option; exclude metadata dir by default
Prevents false mismatches with e.g. md5-cache updates on client side;
option is ignored when verifying, since a flag specifying whether or
not the metadata directory was considered at create time is written
to the repo.hash file. If this flag is absent, metadata is assumed
to be included in the hash, which is backwards-compatible behaviour.